pyproject.toml
[project]
name = "langchecker-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "agno>=2.4.8",
    "openai>=2.17.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=21.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "pytest>=8.4.2",
    "ruff>=0.15.0",
]
